Refactor: Archive old test files
