+++ programs/tests/archives/proposal.test.ts
@@ -4,7 +4,7 @@ console.log("=====================================\n");
 
 import * as anchor from "@coral-xyz/anchor";
 import { Program } from "@coral-xyz/anchor";
-import { Programs } from "../target/types/programs";
+import { Programs } from "../../target/types/programs";
 import { PublicKey, Keypair, SystemProgram, LAMPORTS_PER_SOL } from "@solana/web3.js";
 
 // Importer et configurer chai AVANT d'importer expect
@@ -13,7 +13,7 @@ import chai from "chai";
 // Importer expect de chai APRÈS la configuration
 import { expect } from "chai";
 
-import { generateRandomId, setupTestEnvironment } from "./utils.test";
+import { generateRandomId, setupTestEnvironment } from "../utils.test";
 
 // Seed fixe pour l'autorité admin des tests
 const ADMIN_SEED = Uint8Array.from([1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32]);
